requirements.txt
django==3.1
djangorestframework==3.11.0
django-rest-auth==0.9.3
django-allauth==0.39.1